+++ code/headers/Color_Buffer_Rgb565.hpp
@@ -50,7 +50,7 @@
 
         public:
 
-            Color_Buffer_Rgb565(size_t width, size_t height)
+            Color_Buffer_Rgb565(int width, int height)
             :
                 Color_Buffer(width,  height),
                 buffer      (width * height)
@@ -74,7 +74,7 @@
                 return (sizeof(Color) * 8);
             }
 
-            size_t size () const
+            int size () const
             {
                 return (buffer.size ());
             }
@@ -96,7 +96,7 @@
                 buffer[y * width + x] = color;
             }
 
-            void set_pixel (size_t offset)
+            void set_pixel (int offset)
             {
                 buffer[offset] = color;
             }
